requirements.txt
wheel==0.45.1
opencv-python==4.8.1.78
numpy==1.24.3
pyautogui==0.9.54
pandas==2.2.3
setuptools==78.1.0
cmake==3.27.7
imutils==0.5.4
pytesseract==0.3.10
Pillow==10.0.0
pyperclip==1.8.2
gTTS==2.5.1
playsound==1.3.0
MouseInfo==0.1.3
PyAutoGUI==0.9.54
pybind11==2.13.6
PyGetWindow==0.0.9
matplotlib>=3.8
PyMsgBox==1.0.9
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
tk==0.1.0
dlib==19.24.6
imutils==0.5.4
MouseInfo==0.1.3
keyboard==0.13.5